requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.26.0
gunicorn==21.2.0